backend/requirements.txt
accelerate==1.12.0
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
aistudio-sdk==0.3.8
alembic==1.18.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astor==0.8.1
async-timeout==4.0.3
asyncpg==0.31.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
bce-python-sdk==0.9.60
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.40.61
botocore==1.40.61
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.4
cssselect==1.4.0
cssutils==2.11.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==3.2.4
dataclasses-json==0.6.7
decorator==5.2.1
diffusers==0.36.0
distro==1.9.0
ecdsa==0.19.1
editdistance==0.8.1
einops==0.8.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
funasr==1.3.1
future==1.0.0

GPUtil==1.4.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.5
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kaldiio==2.18.1
kiwisolver==1.4.9
langchain>=0.3.27
langchain-classic==1.0.1
langchain-community>=0.2.4
langchain-core>=0.2.43
langchain-openai==1.1.7
langchain-text-splitters>=0.3.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.1.147
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lmdb==1.7.5
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.16.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.6
ormsgpack==1.12.2
oss2==2.19.1
packaging==23.2
paddleocr==2.6.1.3
paddlepaddle-gpu==2.5.2

pandas==2.3.3
pdf2docx==0.5.8
pillow==12.1.0
platformdirs==4.5.1
polars==1.37.1
polars-runtime-32==1.37.1
pooch==1.9.0
premailer==3.10.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.20.2
pynndescent==0.6.0
pyparsing==3.3.2
pypdfium2==5.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.14
RapidFuzz==3.14.3
rarfile==4.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
ruamel.yaml==0.19.1
s3transfer==0.14.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tenacity==8.5.0
tensorboardX==2.6.4
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torch-complex==0.4.4
torchaudio==2.10.0
torchvision==0.25.0
tqdm>=4.66.3
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
ultralytics==8.4.9
ultralytics-thop==2.0.18
umap-learn==0.5.11
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
visualdl==2.5.3
watchfiles==1.1.1
wcwidth==0.5.3
Werkzeug==3.1.5
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
pypandoc
easyofd==0.5.6
pikepdf==9.0.0
reportlab==4.2.0
svglib==1.5.1
xhtml2pdf>=0.2.17
pdf2docx==0.5.8
playwright==1.58.0
scrapy
moviepy
